add ticket page; add icon; clean up pages
